requirements.txt
Flask
flask-cors
langchain
langchain-community
langchain-huggingface
langchain-core
python-dotenv
faiss-cpu
sentence-transformers
transformers
tokenizers
torch
pypdf
huggingface-hub
gunicorn
numpy
requests
urllib3